Improve socketio error handling behaivor
