+++ querybook/server/const/data_doc.py
@@ -6,3 +6,6 @@ class DataCellType(Enum):
     query = 0
     text = 1
     chart = 2
+
+
+DATA_DOC_NAMESPACE = "/datadoc"
