+++ querybook/server/datasources_socketio/__init__.py
@@ -1,5 +1,7 @@
 from . import query_execution
 from . import datadoc
+from . import connect
 
+connect
 query_execution
 datadoc
